Implemented API error codes and REST collection request modes (fixes #11).
Regarding REST collection requests, for example `/users/`, `/tags/` and so on the default for the system is "strict".
This means submitting and getting data from collection URLs just work with "iterables".

This behavior can be changed when REST does not need to be enforced because of frontend needs.

There are 2 ways of changing the behavior to allow, for example to POST single JSON object and get a single object back.
One is by setting `request.rest_collection_mode = permissive` in settings file.
The other is to use an HTTP request header like `X-REST-Collection-Mode: permissive`.